requirements.txt
cffi==1.15.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
Pillow==9.3.0
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrtlsdr==0.2.93
python-dateutil==2.8.2
scipy==1.9.3
six==1.16.0
sounddevice==0.4.5
requests==2.28.2
websockets
aiohttp
asyncio
numpy
pyrtlsdr==0.2.93